+++ src/attendee.js
@@ -1,5 +1,11 @@
 Occasion.Modules.push(function(library) {
-  library.Attendee = class Attendee extends library.Base {};
+  library.Attendee = class Attendee extends library.Base {
+    complete() {
+      return !this.order().product().attendeeQuestions.detect((question) => {
+        return !this[question] || this[question].length == 0;
+      });
+    }
+  };
 
   library.Attendee.className = 'Attendee';
   library.Attendee.queryName = 'attendees';